+++ Manager/Desktop/Desktop.csproj
@@ -14,7 +14,10 @@
 
   <ItemGroup>
     <PackageReference Include="GtkSharp" Version="3.24.24.*" />
-    <PackageReference Include="TagLibSharp" Version="2.3.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Manager.Implementation\Manager.Implementation.csproj" />
   </ItemGroup>
 
 </Project>
