+++ Manager/Manager.Implementation/Resource/Resource.cs
@@ -1,4 +1,6 @@
-namespace Manager.Resource.Implementation;
+using Manager.Resource;
+
+namespace Manager.Implementation.Resource;
 
 internal class Resource(string identifier, string name) : IResource
 {
