+++ Manager/Manager.Tests/Manager.Tests.csproj
@@ -20,6 +20,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Manager.Implementation\Manager.Implementation.csproj" />
     <ProjectReference Include="..\Manager\Manager.csproj" />
   </ItemGroup>
