+++ Manager/MusicTagsManager.sln
@@ -6,6 +6,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Manager.Tests", "Manager.Te
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Desktop", "Desktop\Desktop.csproj", "{6EF432A9-5944-4E8D-B2B0-B47397C0E40F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Manager.Implementation", "Manager.Implementation\Manager.Implementation.csproj", "{6E9F8D15-F3F6-4112-97FD-52A68499A0B9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -52,6 +54,18 @@ Global
 		{6EF432A9-5944-4E8D-B2B0-B47397C0E40F}.Release|x64.Build.0 = Release|Any CPU
 		{6EF432A9-5944-4E8D-B2B0-B47397C0E40F}.Release|x86.ActiveCfg = Release|Any CPU
 		{6EF432A9-5944-4E8D-B2B0-B47397C0E40F}.Release|x86.Build.0 = Release|Any CPU
+		{6E9F8D15-F3F6-4112-97FD-52A68499A0B9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6E9F8D15-F3F6-4112-97FD-52A68499A0B9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6E9F8D15-F3F6-4112-97FD-52A68499A0B9}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6E9F8D15-F3F6-4112-97FD-52A68499A0B9}.Debug|x64.Build.0 = Debug|Any CPU
+		{6E9F8D15-F3F6-4112-97FD-52A68499A0B9}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6E9F8D15-F3F6-4112-97FD-52A68499A0B9}.Debug|x86.Build.0 = Debug|Any CPU
+		{6E9F8D15-F3F6-4112-97FD-52A68499A0B9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6E9F8D15-F3F6-4112-97FD-52A68499A0B9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6E9F8D15-F3F6-4112-97FD-52A68499A0B9}.Release|x64.ActiveCfg = Release|Any CPU
+		{6E9F8D15-F3F6-4112-97FD-52A68499A0B9}.Release|x64.Build.0 = Release|Any CPU
+		{6E9F8D15-F3F6-4112-97FD-52A68499A0B9}.Release|x86.ActiveCfg = Release|Any CPU
+		{6E9F8D15-F3F6-4112-97FD-52A68499A0B9}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
